+++ bin/templates/version/page-directive.js
@@ -1,9 +1,0 @@
-window.page.directive = {
-    "#version span": "version",
-    "#author span": "author",
-    "#sitename span": "name",
-    "#description span": "description",
-    "#license span": "license",
-    "#code@href": "repository.url",
-    "#bugs@href": "bugs.url",
-}