+++ History.md
@@ -9,3 +9,9 @@
   * Refactor of codebase to fit the scaffolding of an express site
   * Includes all of the features that were developed in conjunction with moltensteelman.com
   * Changed the structure of the framework to hold everything specific to the site in the bin folder (except for the configuration)
+
+0.1.1 -refactor / 2017-02-25
+==================
+
+  * More refactor of codebase to fit express scaffolding
+  * Split the structure of the app between; the bin folder, which holds the static data, schemas, and content, and the build folder which contains the output of the build and an index to run the app.